Fixing nullpointers and such

+++ syncany-lib/src/main/java/org/syncany/connection/plugins/local/LocalTransferManager.java
@@ -30,6 +30,7 @@ import java.util.Set;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
+import org.apache.commons.io.FileExistsException;
 import org.apache.commons.io.FileUtils;
 import org.syncany.connection.plugins.AbstractTransferManager;
 import org.syncany.connection.plugins.DatabaseRemoteFile;
@@ -178,7 +179,21 @@ public class LocalTransferManager extends AbstractTransferManager {
 		File target = getRemoteFile(targetFile);
 
 		try {
-			FileUtils.moveFile(source, target);
+			if (target.exists()) {
+				if (source.length() == target.length()) {
+					//target.delete();
+					//FileUtils.moveFile(source, target);
+				}
+			}
+			else {
+				FileUtils.moveFile(source, target);
+			}
+		}
+		catch (FileExistsException ex) {
+			if (source.length() != target.length()) {
+				target.delete();
+			}
+			
 		}
 		catch (IOException ex) {
 			throw new StorageException("Unable to move file " + source + " to destination " + target, ex);
